Have logic in place to write an array of DatedDatums to an HDF5 file.  Now need to test.
git-svn-id: svn+ssh://svn.oneunified.net/var/local/svn/private/projects/oneunified/projects/trunk/TradeFrame@73 43143c16-083f-4785-a0a0-245a1fb241a9

+++ GTScalp/ChartArmsIntraDay.cpp
@@ -102,7 +102,7 @@ void CChartArmsIntraDay::DrawChart( void ) {
   pChart ->addText( 250,  15, "DnSn!!" );
 
   vector<double>::iterator iterIndu, iterTrin;
-  unsigned int i, j;
+  size_t i, j;
   j = m_vdblIndu.size();
   double *pdblIndu = new double[ j ];
   double *pdblTrin = new double[ j ];
